Introduce `SuspendedCurve` public
This curve is used by bottom sheets and was previously private.

I rewrote the documentation and refactored to be in line with other `Curve` subclasses.
